Regenerate Google Refresh Token
The Google Refresh Token is required for the Acceptance Tests to run
against the live environment and check the sending of emails by
intereacting with the Gmail API.

The Acceptance Tests will refresh an Access Token every run, but
eventually the Refresh Token will need to be replaced. These
instructions show how to do that.